requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
langchain-openai>=0.0.5
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
